fix(сводка синков): нули при рабочих кронах + мусор в cron.log
ПРИЧИНА НУЛЕЙ. Сводка считала прогоны грепом по лог-файлам: 'Starting menu sync', 'DONE duration_ms', 'Updated sync marker'. После перехода кухни и телеграм-алертов на структурное логирование две последние строки писаться перестали. Проверено на проде: telegram.log не обновлялся с 16 мая, 'Updated sync marker' в cron.log нет вовсе — при этом оба крона отрабатывают (kitchen_last_sync_at и telegram_last_run_at свежие). То есть отчёт врал в самую тревожную сторону: показывал 0 там, где всё работало.

ЧТО СДЕЛАНО. Счётчики прогонов переехали в БД (MetaRepository::bumpDailyRun, ключи *_runs_json, окно 7 суток). Не зависят от формата логов и переживают ротацию. Греп оставлен запасным путём — он корректен для меню и выручит первые сутки, пока счётчик пуст.

ЧИТАЕМОСТЬ. Строки отчёта теперь вида «Кухня: ✅ 288/288 за сутки (раз в 5 мин), последний ...»: видно не только факт, но и норму. Меньше 90% — ⚠️, ноль — ❌.

МУСОР В ЛОГЕ. cron.log вырос до 711 МБ и наполовину состоит из 'Deprecated: Implicit conversion from float ... to int' — KitchenAnalytics отдавал float в date(). Добавлен (int); спам прекращается.

+++ src/classes/KitchenAnalytics.php
@@ -225,7 +225,7 @@ class KitchenAnalytics {
                 $seq = 1;
                 foreach ($instances as $inst) {
                     $results[] = [
-                    'date' => date('Y-m-d', ($tx['date_start'] ?? time()*1000) / 1000),
+                    'date' => date('Y-m-d', (int)(($tx['date_start'] ?? time()*1000) / 1000)),
                     'receipt_number' => $tx['receipt_number'] ?? ($tx['transaction_id'] ?? 'N/A'),
                     'transaction_opened_at' => isset($tx['date_start']) ? $this->formatTimestamp($tx['date_start']) : null,
                     'transaction_closed_at' => $this->resolveClosedAt($tx),
